refactor(test): merge compose [id] endpoint tests into route.test.ts (#9394)
Consolidate delete-storage-cleanup.test.ts and permission.test.ts into a
single route.test.ts file under app/api/agent/composes/[id]/__tests__/.
All 7 test cases preserved with no logic changes.
